refactor(controller): move test-only exports behind the test build tag
`internal/adapter/` の本番ファイルに、テストからしか呼ばれないエクスポート
関数が 13 個あった。`internal/domain` には `*_testhelpers.go` +
`//go:build test` の規約が 62 ファイルで確立しているのに、adapter だけ
未追随だった。13 件を規約どおりの位置へ移す。

**issue #5362 に書いた効果のうち 2 つは実測の結果まちがいだったので、
主張を訂正する。**

- ❌ 「deadcode の誤検知が 31 件減る」→ 減らない。実測 13/411 のまま。
  `deadcode -tags test` は test タグ付きファイルも解析対象に含めるので、
  置き場所を変えても main からの到達性は変わらない
- ❌ 「本番バイナリに含まれる」→ 含まれていなかった。develop と本ブランチ
  それぞれのバイナリを `go tool nm` で比較して 0 と 0。参照の無い
  エクスポート関数はリンカが既に除去していた
- ✅ 残る効果は API 衛生のみ。`controller` パッケージの本番ビルドの
  エクスポート面からテスト専用関数が消え、「テスト専用」がコンパイラに
  強制されるようになる（従来は `ForTest` という命名規約への信頼だけ）

build タグは合成 (`test && (!js || !wasm || extra2)`) も検討したが、素の
`test` にした。負のコントロールで確かめたところ、worker+test の組み合わせは
**既存の internal/domain/*_testhelpers.go でも同様に壊れており**、
サポートされたビルドではない。合成タグは 62 ファイルの規約から外れるだけで
新しい保護を生まないと判断した。

検証したビルドは実際に使われる 3 系統すべて:
  go build ./...                      通過
  go vet -tags test ./internal/adapter/...  通過
  6 worker の GOOS=js GOARCH=wasm      すべて通過

Closes #5362
