+++ internal/adapter/controller/PopeJoanWebController.go
@@ -155,9 +155,3 @@ func popeJoanDispatch(bc *baseController, w http.ResponseWriter, pi usecase.Pope
 	}
 	return true
 }
-
-// NewPopeJoanDefaultOutputForTest exposes the default-output builder to the
-// external controller_test package.
-func NewPopeJoanDefaultOutputForTest(msg string) *PopeJoanWebOutput {
-	return newPopeJoanDefaultOutput(msg)
-}
